userverification/serializer.py
from rest_framework import serializers
from .models import UserProfile
from django.contrib.auth.models import User

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ["phone_number"]

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["first_name","last_name","username"]

    def to_representation(self,instance):

        return {"first_name": instance.first_name,
                "last_name": instance.last_name,
                "phone_number": instance.username}