requirement.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.5
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
eskiz-sms==0.2.3
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
multidict==6.0.4
phonenumbers==8.13.26
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
twilio==8.10.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4